@@ -1,52 +1,55 @@
 {
-  "name": "apps",
+  "name": "asm",
   "version": "0.0.0",
   "scripts": {
     "ng": "ng",
-    "start": "ng serve",
+    "start": "node server.js",
+    "postinstall": "ng build",
     "build": "ng build",
-    "watch": "ng build --watch --configuration development",
+    "dev": "ng serve --open --host=0.0.0.0 --port=8080",
     "test": "ng test"
+  },
+  "engines": {
+    "node": "v16.15.1",
+    "npm": "8.11.0"
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "~12.0.5",
-    "@angular/cdk": "^12.0.5",
-    "@angular/common": "~12.0.5",
-    "@angular/compiler": "~12.0.5",
-    "@angular/core": "~12.0.5",
-    "@angular/forms": "~12.0.5",
-    "@angular/material": "^12.0.5",
-    "@angular/platform-browser": "~12.0.5",
-    "@angular/platform-browser-dynamic": "~12.0.5",
-    "@angular/router": "~12.0.5",
-    "@ngrx/effects": "^12.1.0",
-    "@ngrx/entity": "^12.2.0",
-    "@ngrx/router-store": "^12.2.0",
-    "@ngrx/store": "^12.1.0",
-    "@ngrx/store-devtools": "^12.1.0",
-    "@ngx-translate/core": "^13.0.0",
-    "@ngx-translate/http-loader": "^6.0.0",
-    "bootstrap": "^5.0.2",
+    "@angular/animations": "~12.1.0",
+    "@angular/common": "~12.1.0",
+    "@angular/compiler": "~12.1.0",
+    "@angular/core": "~12.1.0",
+    "@angular/forms": "~12.1.0",
+    "@angular/platform-browser": "~12.1.0",
+    "@angular/platform-browser-dynamic": "~12.1.0",
+    "@angular/router": "~12.1.0",
+    "@angular/cdk": "~12.1.0",
+    "@angular/material": "~12.1.0",
+    "@ngrx/effects": "~12.1.0",
+    "@ngrx/entity": "~12.1.0",
+    "@ngrx/router-store": "~12.1.0",
+    "@ngrx/store": "~12.1.0",
+    "@ngrx/store-devtools": "~12.1.0",
+    "bootstrap": "^5.1.3",
     "browser-detect": "^0.2.28",
-    "ngx-currency": "^2.5.2",
-    "ngx-device-detector": "^2.1.1",
+    "express": "^4.18.1",
+    "ngx-currency": "^2.5.3",
+    "path": "^0.12.7",
     "rxjs": "~6.6.0",
-    "tslib": "^2.1.0",
+    "tslib": "^2.2.0",
     "zone.js": "~0.11.4"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~12.0.5",
-    "@angular/cli": "~12.0.5",
-    "@angular/compiler-cli": "~12.0.5",
-    "@types/jasmine": "~3.6.0",
-    "@types/node": "^12.11.1",
-    "jasmine-core": "~3.7.0",
+    "@angular-devkit/build-angular": "~12.1.4",
+    "@angular/cli": "~12.1.4",
+    "@angular/compiler-cli": "~12.1.0",
+    "@types/jasmine": "~3.8.0",
+    "jasmine-core": "~3.8.0",
     "karma": "~6.3.0",
     "karma-chrome-launcher": "~3.1.0",
     "karma-coverage": "~2.0.3",
     "karma-jasmine": "~4.0.0",
-    "karma-jasmine-html-reporter": "^1.5.0",
-    "typescript": "~4.2.3"
+    "karma-jasmine-html-reporter": "~1.7.0",
+    "typescript": "~4.3.2"
   }
 }
