@@ -1,8 +1,9 @@
-import { Component, ViewChild } from '@angular/core';
+import { Component, ViewChild, ChangeDetectionStrategy } from '@angular/core';
 import { MatSidenav } from '@angular/material/sidenav';
 import { BreakpointObserver } from '@angular/cdk/layout';
 
 @Component({
+  changeDetection: ChangeDetectionStrategy.OnPush,
   selector: 'app-root',
   templateUrl: './app.component.html',
   styleUrls: ['./app.component.scss']
@@ -17,9 +18,13 @@ export class AppComponent {
 
   constructor(private observer: BreakpointObserver) {}
 
-  ngOnInit() {}
+  ngOnInit() {
 
-  ngAfterViewInit() {
+    /**
+     * Se a aba do navegador for arrastada para tamnho que
+     * o menu lateral irar sumir
+     * se for arrastada para maior o menu lateral ira aparecer
+     */
     this.observer.observe(['(max-width: 800px)']).subscribe((res) => {
       if (res.matches) {
         this.sidenav.mode = 'over';
@@ -29,5 +34,17 @@ export class AppComponent {
         this.sidenav.open();
       }
     });
+  }
+
+  /**
+   * em modo mobile, quando o menu lateral for clicado, oculta e mostra a tela referente aquele menu
+   */
+  public sideNavClose(){
+    this.observer.observe(['(max-width: 800px)']).subscribe((res) => {
+      if (res.matches) {
+        this.sidenav.mode = 'over';
+        this.sidenav.close();
+      }
+    })
   }
 }
